demo-agent/examples: sweep the callers the hand-STARK retirement missed (+ close the door)
demo-agent/examples/ was broken in 6 examples. The predicate/threshold failures
were NOT independent unsweeps — 11 unresolved symbols across 4 examples are ONE
architectural move whose callers were never swept:

  dregg-circuit used to own monolithic hand-STARK prove_X/verify_X. That engine
  was retired. dregg-circuit now owns descriptors + witness generators, and
  dregg-bridge::present owns prove/verify over the IR-v2 descriptor prover.

The library crates WERE swept — wasm/src/lib.rs:1446, sdk/src/verify.rs:719,
sdk/src/cipherclerk.rs:4420 and bridge/src/present.rs:3092 all carry deliberate
fail-closed stubs with comments explaining the retirement. demo-agent/examples/
was the only unswept caller set, because nothing compiled it.

The remaining two (compute_marketplace, delegation_swarm) are unrelated single-
signature drifts, and are described below — one of them is NOT mechanical.

WHICH CAPABILITIES SURVIVED (registered in descriptor_by_name, really provable):
  - Gte / Lte / Gt / Lt / Neq / InRange  via prove_predicate_for_fact*
  - attested fact membership (dregg-attested-fact-membership::v1)

WHICH WERE REMOVED (no emitted descriptor; entry points fail closed):
  - committed threshold (hidden value AND hidden threshold) —
    bridge::present::prove_committed_threshold returns None unconditionally,
    verify_committed_threshold_proof returns false unconditionally
  - in-circuit compound predicate — no prove fn at all; the descriptor exists but
    is NOT registered, so nothing can produce or check it
  - garbled-evaluation STARK (prove_private_threshold_dsl)

PER EXAMPLE:

* anonymous_credit_check — PORTED. Its premise (hidden threshold) was retired, so
  the threshold is now a PUBLIC input and the module header says so plainly. Moved
  to the THIRD-PARTY shape (prove_predicate_for_fact_attested /
  verify_predicate_proof_third_party), not the trusted-state shape: the bank does
  not know Alice's score, so it cannot re-derive her fact commitment, and the
  trusted-state entry point is sound only for a verifier that already holds the
  value. The bureau's facts_root is derived independently via attested_facts_root,
  never read off the prover's own attestation. Four adversarial poles, each with a
  neutered canary where one applies.

* private_auction — PORTED. Bid-validity predicates moved to the attested shape.
  The committed-threshold winner-determination phase was DELETED rather than
  rewired to a fail-closed call that would only print failure; prose corrected to
  match what the code reveals. Notes / NullifierSet / conditional settlement
  untouched.

* private_hiring — PORTED. Four predicates to the attested shape (the skill check
  is now Predicate::Neq(0), directly expressible where the old code used Gte(1) as
  a workaround). Phase 6's "single compound proof" is gone: the AND is now
  evaluated VERIFIER-SIDE by evaluate_formula over the actual return values of four
  independent verifications, with a non-vacuity check that flipping one conjunct
  makes the AND reject. The old code fed it vec![true; 4], which proved nothing.

* garbled_ot_auction — REWRITTEN around the surviving capability. The real 2PC
  (Chou-Orlandi OT + Yao garbled evaluation) is intact and is what the demo now
  shows; the retired STARK-of-correct-evaluation is named, along with the honest
  consequence (integrity now rests on the garbler decoding, not on a proof a third
  party can check). Added the assertion the demo never had: the garbled outcome
  must equal the plaintext comparison, and the tournament must pick the true
  maximum.

* compute_marketplace — FIXED, and NOT mechanically. NullifierSet::insert gained a
  second parameter: the spent note's u64 value, which it folds into the grow-gate
  leaf to keep root8 byte-identical to the in-circuit noteSpend accumulator. This
  demo was pushing sealed BID COMMITMENTS into that set. Passing the bid — the
  mechanical repair — would have handed the marketplace every bid at COMMIT time,
  destroying the exact property the phase exists to show; passing 0 would write a
  value the accumulator's docs say must be real. A sealed bid is not a spent note,
  so the commit-once registry is now an ordinary BTreeSet over the commitments and
  the note accumulator is left to notes. The three inserts now assert `true` so the
  double-submit `false` is a refusal rather than a set that rejects everything.

* delegation_swarm — FIXED (genuinely mechanical): Effect::SetField::index is u64
  (it is hashed via to_le_bytes into the action commitment) and the loop counter is
  usize.

* base_anonymous_credential — DELETED. Built on the retired committed-threshold
  AIR, and its headline claim (anonymous credential verified ON-CHAIN via SP1
  STARK-to-Groth16 on Base) was entirely simulated: simulate_onchain_verify
  returned true for ANY non-empty byte slice, the "Groth16 proof" was a blake3
  hash, and ring membership was a printed narrative. Nothing in it executed. The
  real surviving capability is demonstrated end-to-end by anonymous_credit_check.

CLOSING THE DOOR. The gate already existed: ci.yml:49 runs
cargo check --workspace --all-targets, which does cover these targets. It never
caught this because it is a whole-workspace job that dies on the first crate to
break — until 1db1cd3171 (yesterday) it died at dregg-circuit on two missing
descriptor JSONs — and CI on main is chronically failure/cancelled, so a new
example break was indistinguishable from standing noise. Rather than add a
redundant gate that would be masked the same way, demos.yml's existing build step
now says --bins --examples: a narrow, legible signal in a job that already builds
this crate. (--bins is not redundant — cargo target selection is restrictive, so a
bare --examples would stop producing the binary the next step runs.) Cost: near
zero, every dependency is already compiled by that step; --examples adds only
per-example codegen, ~1-3 min warm.

Also fixed three STALE gate entries that could only ever have failed: demos.yml's
run list named federation_bootstrap, offline_verification and wallet_lifecycle,
none of which have [[example]] entries; preflight's list named
base_anonymous_credential, bench_summary and offline_verification. Manifest, disk,
and preflight are now in sync (38 examples, zero orphans in either direction).

THIS COMMIT ALSO UN-BREAKS HEAD. The base_anonymous_credential.rs deletion was
staged here and then swept into an unrelated commit by another lane
(555f044420, "pbuild: forward the DREGG_* knobs to the remote") — the git add -A
hazard. That commit took the FILE but not the manifest entry, so HEAD has an
[[example]] stanza pointing at a file that does not exist and dregg-demo-agent
does not build there. Dropping the stanza here repairs it, and the deletion's
rationale — which the sweeping commit's message does not contain — is recorded
both above and as a durable comment in demo-agent/Cargo.toml.

NOTE ON THIS COMMIT: demo-agent/Cargo.toml also carries a dregg-turn-prover dep
line that was already in the worktree from another lane (for cross_fed_atomic);
git commit --only on the manifest could not exclude it, and the manifest had to
change to drop the deleted example.

+++ demo-agent/examples/delegation_swarm.rs
@@ -248,7 +248,9 @@ fn main() {
                 preconditions: Default::default(),
                 effects: vec![Effect::SetField {
                     cell: target_svc,
-                    index: turn_idx % 8,
+                    // `Effect::SetField::index` is a `u64` (it is hashed via `to_le_bytes` into the
+                    // action commitment); `turn_idx` is a loop `usize`.
+                    index: (turn_idx % 8) as u64,
                     value,
                 }],
                 may_delegate: DelegationMode::None,
